test(enhancement 3): Trader Dashboard and Blotter System - Added unit tests in the dashboardServiceTest and added error handling to handle when the user has no data for the dashboard views.

+++ backend/src/main/java/com/technicalchallenge/dto/TradeSummaryDTO.java
@@ -71,14 +71,14 @@ public class TradeSummaryDTO {
         // Flatten Version of breakdownByTradeType
         public static record TradeTypeBreakdown(
                         String tradeType,
-                        BigDecimal totalNotional, BigDecimal percentage) {
+                        BigDecimal totalNotional, Double percentage) {
         }
 
         // Flatten Version of breakdownByCounterparty
         public static record CounterpartyBreakdown(
                         String counterpartyName,
                         BigDecimal totalNotional,
-                        BigDecimal percentage) {
+                        Double percentage) {
 
         }
